scraper/requirements.txt
Scrapy
psycopg2-binary
python-dotenv